valores objetivos

+++ database/migrations/2021_10_30_133713_create_objetivos_table.php
@@ -17,7 +17,7 @@ class CreateObjetivosTable extends Migration
             $table->increments('id');
             $table->string('nome');
             $table->text('descricao')->nullable();
-            $table->decimal('valor');
+            $table->integer('valor');
             $table->date('data_inicial');
             $table->date('data_final');
